Update tests and src file structure
