+++ test/delete.spec.ts
@@ -1,9 +1,10 @@
 // @ts-nocheck
 /* global describe beforeAll beforeEach afterEach it expect */
 
+import url from 'url';
 import nock from 'nock';
 import fetch from 'node-fetch';
-import trae from '../../src';
+import trae from '../src';
 import util from './util';
 import http from 'http';
 
@@ -12,7 +13,7 @@ global.Headers = fetch.Headers;
 
 const TEST_URL = 'http://localhost:8080';
 
-describe('trae -> post', () => {
+describe('trae -> delete', () => {
   describe('Using nock', () => {
     let request;
     let response;
@@ -23,48 +24,40 @@ describe('trae -> post', () => {
           'content-type': 'application/json',
         },
       })
-        .post('/foo', { pizza: 'guerrin' })
-        .reply(200, { foo: 'bar' });
+        .delete('/airports/barcelona')
+        .reply(204);
     });
 
     beforeAll(function executeRequest() {
-      return trae
-        .post(TEST_URL + '/foo', { pizza: 'guerrin' })
-        .then(function(res) {
-          response = res;
-        });
+      return trae.delete(TEST_URL + '/airports/barcelona').then(function (res) {
+        response = res;
+      });
     });
 
-    it('should make an HTTP post request', function() {
+    it('should make an HTTP delete request', function () {
       const actual = request.isDone();
       const expected = true;
 
       expect(actual).toStrictEqual(expected);
     });
 
-    it('should have 200 in the response status code', function() {
+    it('should have 204 in the response status code', function () {
       const actual = response.status;
-      const expected = 200;
+      const expected = 204;
 
       expect(actual).toStrictEqual(expected);
     });
 
-    it('should have "OK" in the response status text', function() {
+    it('should have "No Content" in the response status text', function () {
       const actual = response.statusText;
-      const expected = 'OK';
+      const expected = 'No Content';
 
       expect(actual).toStrictEqual(expected);
     });
 
-    it('should have foo bar in the response data', function() {
-      const actual = response.data;
-      const expected = { foo: 'bar' };
-
-      expect(actual).toStrictEqual(expected);
-    });
-
-    describe('post using params', function() {
+    describe('delete using params', function () {
       let request;
+      let response;
 
       beforeAll(function createNock() {
         request = nock(TEST_URL, {
@@ -72,33 +65,37 @@ describe('trae -> post', () => {
             'content-type': 'application/json',
           },
         })
-          .post('/cats?name=tigrin')
+          .delete('/baz')
+          .query({ name: 'foo' })
           .reply(200, { yay: 'OK' });
       });
 
       beforeAll(function executeRequest() {
-        return trae.post(
-          TEST_URL + '/cats',
-          {},
-          { params: { name: 'tigrin' } },
-        );
+        return trae.delete(TEST_URL + '/baz', {
+          params: {
+            name: 'foo',
+          },
+        })
+          .then(function (res) {
+            response = res
+          });
       });
 
-      it('should make an HTTP post request', function() {
+      it('should make an HTTP delete request', function () {
         const actual = request.isDone();
         const expected = true;
 
         expect(actual).toStrictEqual(expected);
       });
 
-      it('should have 200 in the response status code', function() {
+      it('should have 200 in the response status code', function () {
         const actual = response.status;
         const expected = 200;
 
         expect(actual).toStrictEqual(expected);
       });
 
-      it('should have "OK" in the response status text', function() {
+      it('should have "OK" in the response status text', function () {
         const actual = response.statusText;
         const expected = 'OK';
 
@@ -106,8 +103,9 @@ describe('trae -> post', () => {
       });
     });
 
-    describe('post using nested params', function() {
+    describe('delete using nested params', function () {
       let request;
+      let response;
 
       beforeAll(function createNock() {
         request = nock(TEST_URL, {
@@ -115,39 +113,38 @@ describe('trae -> post', () => {
             'content-type': 'application/json',
           },
         })
-          .post('/foo?a%5Bb%5D=c')
+          .delete('/foo?a%5Bb%5D=c')
           .reply(200);
       });
 
       beforeAll(function executeRequest() {
-        return trae.post(
-          TEST_URL + '/foo',
-          {},
-          {
-            params: {
-              a: {
-                b: 'c',
-              },
+        return trae.delete(TEST_URL + '/foo', {
+          params: {
+            a: {
+              b: 'c',
             },
           },
-        );
+        })
+          .then(function (res) {
+            response = res
+          })
       });
 
-      it('should make an HTTP post request', function() {
+      it('should make an HTTP delete request', function () {
         const actual = request.isDone();
         const expected = true;
 
         expect(actual).toStrictEqual(expected);
       });
 
-      it('should have 200 in the response status code', function() {
+      it('should have 200 in the response status code', function () {
         const actual = response.status;
         const expected = 200;
 
         expect(actual).toStrictEqual(expected);
       });
 
-      it('should have "OK" in the response status text', function() {
+      it('should have "OK" in the response status text', function () {
         const actual = response.statusText;
         const expected = 'OK';
 
@@ -162,52 +159,46 @@ describe('trae -> post', () => {
 
     beforeAll(function createServer() {
       function handler(req: http.IncomingMessage, res: http.ServerResponse) {
-        res.setHeader('Content-Type', 'application/json');
-        res.end(JSON.stringify(req.body));
+        res.writeHead(204);
+        res.end();
       }
 
       server = util.createServer({
-        port: 8085,
-        endpoint: '/cities/echo',
+        port: 8081,
+        endpoint: '/delete',
         handler,
       });
     });
 
     beforeAll(function executeRequest() {
-      return trae
-        .post('http://localhost:8085/cities/echo', {
-          city: 'istanbul',
-        })
-        .then(function(res) {
-          response = res;
-        });
+      return trae.delete('http://localhost:8081' + '/delete').then(function (res) {
+        response = res;
+      });
     });
 
     afterAll((done) => server.shutdown(done));
 
-    it('should make an HTTP post request', function() {
+    it('should make an HTTP delete request', function () {
       expect(response).toBeDefined();
     });
 
-    it('should have the defined response data', function() {
+    it('should not have data in the response', function () {
       const actual = response.data;
-      const expected = {
-        city: 'istanbul',
-      };
+      const expected = '';
 
       expect(actual).toStrictEqual(expected);
     });
 
-    it('should have 200 in the response status code', function() {
+    it('should have 204 in the response status code', function () {
       const actual = response.status;
-      const expected = 200;
+      const expected = 204;
 
       expect(actual).toStrictEqual(expected);
     });
 
-    it('should have "OK" in the response status text', function() {
+    it('should have "No Content" in the response status text', function () {
       const actual = response.statusText;
-      const expected = 'OK';
+      const expected = 'No Content';
 
       expect(actual).toStrictEqual(expected);
     });
