+++ test/middleware.spec.ts
@@ -2,7 +2,7 @@
 /* global describe it expect afterEach */
 
 import fetchMock from 'fetch-mock';
-import trae      from '../../src';
+import trae      from '../src';
 
 afterEach(() => {
   fetchMock.restore();
