+++ test/post.spec.ts
@@ -3,7 +3,7 @@
 
 import nock from 'nock';
 import fetch from 'node-fetch';
-import trae from '../../src';
+import trae from '../src';
 import util from './util';
 import http from 'http';
 
@@ -12,7 +12,7 @@ global.Headers = fetch.Headers;
 
 const TEST_URL = 'http://localhost:8080';
 
-describe('trae -> patch', () => {
+describe('trae -> post', () => {
   describe('Using nock', () => {
     let request;
     let response;
@@ -23,47 +23,47 @@ describe('trae -> patch', () => {
           'content-type': 'application/json',
         },
       })
-        .patch('/foo', { pizza: 'guerrin' })
+        .post('/foo', { pizza: 'guerrin' })
         .reply(200, { foo: 'bar' });
     });
 
     beforeAll(function executeRequest() {
       return trae
-        .patch(TEST_URL + '/foo', { pizza: 'guerrin' })
-        .then(function (res) {
+        .post(TEST_URL + '/foo', { pizza: 'guerrin' })
+        .then(function(res) {
           response = res;
         });
     });
 
-    it('should make an HTTP patch request', function () {
+    it('should make an HTTP post request', function() {
       const actual = request.isDone();
       const expected = true;
 
       expect(actual).toStrictEqual(expected);
     });
 
-    it('should have 200 in the response status code', function () {
+    it('should have 200 in the response status code', function() {
       const actual = response.status;
       const expected = 200;
 
       expect(actual).toStrictEqual(expected);
     });
 
-    it('should have "OK" in the response status text', function () {
+    it('should have "OK" in the response status text', function() {
       const actual = response.statusText;
       const expected = 'OK';
 
       expect(actual).toStrictEqual(expected);
     });
 
-    it('should have foo bar in the response data', function () {
+    it('should have foo bar in the response data', function() {
       const actual = response.data;
       const expected = { foo: 'bar' };
 
       expect(actual).toStrictEqual(expected);
     });
 
-    describe('patch using params', function () {
+    describe('post using params', function() {
       let request;
 
       beforeAll(function createNock() {
@@ -72,33 +72,33 @@ describe('trae -> patch', () => {
             'content-type': 'application/json',
           },
         })
-          .patch('/cats?name=tigrin')
+          .post('/cats?name=tigrin')
           .reply(200, { yay: 'OK' });
       });
 
       beforeAll(function executeRequest() {
-        return trae.patch(
+        return trae.post(
           TEST_URL + '/cats',
           {},
           { params: { name: 'tigrin' } },
         );
       });
 
-      it('should make an HTTP patch request', function () {
+      it('should make an HTTP post request', function() {
         const actual = request.isDone();
         const expected = true;
 
         expect(actual).toStrictEqual(expected);
       });
 
-      it('should have 200 in the response status code', function () {
+      it('should have 200 in the response status code', function() {
         const actual = response.status;
         const expected = 200;
 
         expect(actual).toStrictEqual(expected);
       });
 
-      it('should have "OK" in the response status text', function () {
+      it('should have "OK" in the response status text', function() {
         const actual = response.statusText;
         const expected = 'OK';
 
@@ -106,7 +106,7 @@ describe('trae -> patch', () => {
       });
     });
 
-    describe('patch using nested params', function () {
+    describe('post using nested params', function() {
       let request;
 
       beforeAll(function createNock() {
@@ -115,12 +115,12 @@ describe('trae -> patch', () => {
             'content-type': 'application/json',
           },
         })
-          .patch('/foo?a%5Bb%5D=c')
+          .post('/foo?a%5Bb%5D=c')
           .reply(200);
       });
 
       beforeAll(function executeRequest() {
-        return trae.patch(
+        return trae.post(
           TEST_URL + '/foo',
           {},
           {
@@ -133,21 +133,21 @@ describe('trae -> patch', () => {
         );
       });
 
-      it('should make an HTTP patch request', function () {
+      it('should make an HTTP post request', function() {
         const actual = request.isDone();
         const expected = true;
 
         expect(actual).toStrictEqual(expected);
       });
 
-      it('should have 200 in the response status code', function () {
+      it('should have 200 in the response status code', function() {
         const actual = response.status;
         const expected = 200;
 
         expect(actual).toStrictEqual(expected);
       });
 
-      it('should have "OK" in the response status text', function () {
+      it('should have "OK" in the response status text', function() {
         const actual = response.statusText;
         const expected = 'OK';
 
@@ -162,12 +162,12 @@ describe('trae -> patch', () => {
 
     beforeAll(function createServer() {
       function handler(req: http.IncomingMessage, res: http.ServerResponse) {
-        res.writeHead(204);
-        res.end();
+        res.setHeader('Content-Type', 'application/json');
+        res.end(JSON.stringify(req.body));
       }
 
       server = util.createServer({
-        port: 8084,
+        port: 8085,
         endpoint: '/cities/echo',
         handler,
       });
@@ -175,37 +175,39 @@ describe('trae -> patch', () => {
 
     beforeAll(function executeRequest() {
       return trae
-        .patch('http://localhost:8084/cities/echo', {
+        .post('http://localhost:8085/cities/echo', {
           city: 'istanbul',
         })
-        .then(function (res) {
+        .then(function(res) {
           response = res;
         });
     });
 
     afterAll((done) => server.shutdown(done));
 
-    it('should make an HTTP patch request', function () {
+    it('should make an HTTP post request', function() {
       expect(response).toBeDefined();
     });
 
-    it('should have no data the response', function () {
+    it('should have the defined response data', function() {
       const actual = response.data;
-      const expected = ""
+      const expected = {
+        city: 'istanbul',
+      };
 
       expect(actual).toStrictEqual(expected);
     });
 
-    it('should have 204 in the response status code', function () {
+    it('should have 200 in the response status code', function() {
       const actual = response.status;
-      const expected = 204;
+      const expected = 200;
 
       expect(actual).toStrictEqual(expected);
     });
 
-    it('should have "No Content" in the response status text', function () {
+    it('should have "OK" in the response status text', function() {
       const actual = response.statusText;
-      const expected = 'No Content';
+      const expected = 'OK';
 
       expect(actual).toStrictEqual(expected);
     });
